⭐️ Container edge releases (#645)
Since we cannot use goreleaser directly to do our edge releases I
created a separate workflow for that. The workflow is triggered for
every commit to `main`. It will do a snapshot release of the binary and
then create a container with that binary. The container is then pushed
to our registry. This PR will create a container for every architecture.
If we agree to the approach I will create a follow-up PR to do the
virtual tag that combines all architectures.

Signed-off-by: Ivan Milchev <ivan@mondoo.com>