🧹 Use proper tag naming for edge releases (#660)
The goal here is to have the following tags for edge releases:
- edge-latest
- edge-latest-rootless
- edge-VERSION
- edge-VERSION-rootless

I spent quite some time trying to make the magic happen with `yq` but it
seems too much effort. 80% of the goreleaser config for edge releases is
the docker configuration. In this case it has to be quite different from
the one for the normal release procedure. This is the reason why I
decided to just manually configure the file and use that

Signed-off-by: Ivan Milchev <ivan@mondoo.com>