🐛 Fix broken documentation when using `aliases`
When defining aliases, it was possible that the alias aliased the same
physical memory. When this happened, the non alias resource definition
would change the comment and remove the leading `//` when calling
`resourceSchema`. If this function got called again, the leading `//`
would not be there the second time, leading to 2 other characters being
chopped off.