🐛 Use consistent k8s resource version label (#848)
Fixes #647

The double labeling described in #647 is present only for k8s admission
controller scans. That is because the admission controller labels
objects with `k8s.mondoo.com/resource-version` and cnquery does that
with `k8s.mondoo.com/resourceVersion`. Since we parse
`k8s.mondoo.com/resource-version` to `Resource Version` in the UI I
picked that one to be the label for resource version. Making it
consistent between cnquery and the operator means that there will be
just 1 label on the scanned assets

Signed-off-by: Ivan Milchev <ivan@mondoo.com>