🐛 fix mql hangs with aliased and direct resource in the same policy (#850)
Example policy that causes things to hang
```
policies:
  - uid: testpolicy3
    name: Test Policy 3
    version: 0.10.1
    is_public: false
    scoring_system: 2
    tags:
      stability: "testing"
    specs:
      - asset_filter:
          query: |
            true == true
        scoring_queries:
          jay-scoring-query-2:
          jay-scoring-query-3:
queries:
  - uid: jay-scoring-query-2
    title: test it 1
    severity: 10
    query: |
      file("foo").content
  - uid: jay-scoring-query-3
    title: test it 2
    severity: 10
    query: |
      os.base.file("foo").content
```

The issue lies in the fact that the cache key used by some of the
observer code differs from the runtime code that caches the resources.
The observer thinks its already seen a resource, so it must already be
computed (or will be computed). However, the runtime code cached it
using the aliased name. It then tries to ask the resources cache to look
up a field, which is empty because its basically treated as a different
resource that needs to be computed.