🧹 Resolve secret for each provider config in GCP (#867)
Initially, the idea was to copy the resolved secret over to the cloned
provider configs. That is not ideal since it requires other changes as
well. After a discussion with @chris-rock we concluded that resolving
secrets multiple times is a vault problem, not a provider problem. This
PR makes sure that scanning GCP assets work. In a follow up PR I will
address the issue with the retrieval of the same secret multiple times.

Signed-off-by: Ivan Milchev <ivan@mondoo.com>