🧹 semantically split up mquery Merge and AddBase (#885)
The current call only allowed the overwriting of the query that was
trying to merge with a base. This is problematic in places where we do
not want to alter the original query. This is not the case in cnquery,
but it is the case in cnspec (which will be updated shortly after this
merged to accomodate the change for v8).

Signed-off-by: Dominik Richter <dominik.richter@gmail.com>