🧹 Unify GCP compute instance discovery with the rest of the GCP asset discovery (#872)
**This change does not affect anything user-facing**

BUT

it does make sure that we no longer discover instances that are not
running. Since we currently do ssh scanning on them it makes no sense to
discover them. This is also what AWS does for EC2 instances. When we
enable cloud scanning + ssh for a single asset we can adjust that logic.

The goal is to remove the code that does instance discovery at the
moment since it doesn't fit the approach that is used for discovering
any of the other GCP assets.

```
./cnquery scan gcp --discover instances 
```

---------

Signed-off-by: Ivan Milchev <ivan@mondoo.com>