ipv6-friendly addresses for darwin/bsd
lsof shows a process listening on any ipv6 address as the address "*".

lsof shows a process listening on a specify ipv6 address as the address "[::1]".

change the listening address for the "*" to be more ipv6/golang-friendly way so that the listening ports presented in the output are more consistent

% ./cnquery run local -c "ports.listening.where(process.executable == 'nginx:')"
→ discover related assets for 1 asset(s)
→ resolved assets resolved-assets=1
ports.listening.where: [
  0: port port=8080 protocol="ipv6" address="[::1]" process.executable="nginx:"
  1: port port=8443 protocol="ipv6" address="[::]" process.executable="nginx:"
  2: port port=8080 protocol="ipv6" address="[::1]" process.executable="nginx:"
  3: port port=8443 protocol="ipv6" address="[::]" process.executable="nginx:"
]

Signed-off-by: Joel Diaz <joel@mondoo.com>