🐛 Make sure GCP alert policies conditions are always parsed correctly (#942)
GCP alert policies condition filters can be complex queries with more
than 1 statement in them. The assumption before was that the condition
filter looks like this:
```
filter: "metric.type=\"logging.googleapis.com/user/major-foo-1\""
```
However, we just learned that this is not the case and valid filters are
also:
```
filter: "metric.type=\"logging.googleapis.com/user/pass-audit-config-changes-0br0\" AND resource.type=\"metric\""
```

This PR makes sure the filter is split on ` ` and then every segment is
checked until we locate the one containing the `metric.type`. Then we
extract the metric name.

Signed-off-by: Ivan Milchev <ivan@mondoo.com>