🐛 fix incorrect enum numbers on actions (#1036)
I had added the current proto on Feb 1st when v8 was started. However,
the numbering seems to be an overview from the AssignmentDelta that I
mixed up. This PR fixes the numbering and needs a follow-up in cnspec
(which will be up momentarily) to revert the enum mapping there.

Signed-off-by: Dominik Richter <dominik.richter@gmail.com>