🧹 final rename: compose => variant
After numerous discussions we have decided to call the field in queries, which controls variants, fittingly "variants"

It's still just query composition as the underlying mechanism, but we
want to use words that make more sense to people. That's why we have
settled on "variant" and "variants" after a lot of feedback!

Signed-off-by: Dominik Richter <dominik.richter@gmail.com>