🧹 Make listing k8s namespaces optional (#1034)
Fixes #1022

This does not make running queries against a cluster for which we have
no namespace rights possible. We need to figure out what to do with the
platformID generation for a cluster. At the moment the UID of the
kube-system namespace is used. However, it is not always the case that
we have access to that

Signed-off-by: Ivan Milchev <ivan@mondoo.com>