Add instances-api azure discovery target (#1057)
Problem: Trying to scan an instance right now with an azure policy does
not work since the instances have an attached SSH connection. This means
that a query like `azure.compute.subscription.vm{...}` will not work on
the ssh provider.

To fix this, this PR introduces a new target `instances-api`.
- `instances` is meant to instantiate the instances as proper VMs with
attached SSH connection to them.
- `instances-api` is meant to instantiate the instances as API objects,
solely looking at what the azure API returns on them

Discovering both will produce 2 assets for 1 VM, one on the 'API'
(external) side and one for the 'OS' (inside) side
```
➜ ~/go/bin/cnquery shell azure --discover instances-api,instances
→ loaded configuration from /Users/preslavgerchev/.config/mondoo/mondoo.yml using source default
→ discover related assets for 1 asset(s)
! could not determine credentials for asset name=ms365-windows
→ resolved assets resolved-assets=2

    Available assets                       
                                           
  > 1. ms365-windows (azure-compute-vm)    
    2. ms365-windows (azure-compute-vm-api)
```

This will allow us to rewrite policies by using the new platform name
for a target so that azure policies only target API instance assets.