Avoid sudo on self-hosted runners (#1073)
Jobs on self hosted runners do not run in containers, they run on the
host. Allowing a workflow to install software or otherwise manipulate
the system as root is a major security risk. Instead lets just check
that it's installed. (Sadly, there isn't an 'if' expression to determine
if we're on a shared or self-hosted runner, which would be ideal.)

Signed-off-by: Ben Rockwood <benr@cuddletech.com>