✨ scoring system unmarshal + marshal yaml/json
Add two methods for unmarshaling scoring system from yaml and json.
These are taken straight from cnspec. A new method has been added to
marshal to yaml, because if we go that route we generate user-facing
content, so it should use words instead of numbers (since the former are
documented, the latter are not).