🐛 ignore empty namespaces for k8s manifests (#1094)
With the changes related to namespace scanning, we now use listing
namespaces in more places. That uncovered a bug in the code that lists
namespaces for a manifest. If a resource in a manifest has an
unspecified namespace, cnquery will return a namespace with no name.
That has a different meaning if k8s. Namespace with no name means "all
namespaces".

This behaviour triggers [this
bug](https://github.com/mondoohq/mondoo-operator/actions/runs/4492994964/jobs/7903820410?pr=727).

Signed-off-by: Ivan Milchev <ivan@mondoo.com>