Unify the wording for each query pack (#57)
We're all over the place with how we talk about the data we're gathering
in cnquery. Most places we say we're "gathering" data so let's use that
everywhere. We don't need to say we're gathering a list or gathering
information since that's just assumed. Just stick to a consistent and
basic form so the output on the CLI looks clean.

---------

Signed-off-by: Tim Smith <tsmith84@gmail.com>
Co-authored-by: Letha <letha@mondoo.com>