🐛 where/contains/... + vars (#1279)
... was incorrectly interpreted as a non-standalone block. That is
wrong. The `standalone` attribute in compilers referes to the contents
of the compiler not to reference its bindings, i.e. everything in the
expressions to stand on their own (or said differently: the expressions
to be standalone). We had a bug where standalone was set to `false` when
variables were used. The use of vars does not make something
non-standalone. Only the values inside of vars can do that, which are
handled in their assignment.

Example:

```
allowed_roles = ["roles/owner", "roles/editor", "roles/viewer"]
role = "roles/owner2"
allowed_roles.contains(role)
```

Since `role` is not in allowed roles, this should fail. Before, this was
interpreted incorrectly and returned true. This bug is now fixed.

---------

Signed-off-by: Dominik Richter <dominik.richter@gmail.com>