🐛 fix piped kubernetes manifests (#1293)
fixes https://github.com/mondoohq/cnspec/issues/622

The initial implementation https://github.com/mondoohq/cnquery/pull/1213
was reading data from stdin directly inside of the provider. This does
not work well for manifest files with multiple entries since the
discovery splits them up into multiple scans. You can read from stdin
only once.

In general I think it is best practice do avoid doing cli / terminal
handling inside of the provider. Therefore this PR moves this logic to
the cli to read the file. The initial version used a temporary file on
disk. Problem with temporary files is that someone has to clean up the
file after everything is complete. In case of a manifest file, multiple
provider instances are created and it would require additional logic to
handle the close down. Instead I opted for the version that reads the
file and attaches the data into the provider options. I think this is
only a temporary solution. Similar to MQL resources, we should introduce
a provider cache where arbitrary data can be stored. Once that exists,
we could use an in-memory file system and share that. The current
implementation is a step in between.

---------

Signed-off-by: Ivan Milchev <ivan@mondoo.com>
Co-authored-by: Ivan Milchev <ivan@mondoo.com>