🏗️ restructure around new providers-sdk
We are moving everything we can from motor and resources over to
providers-sdk. This also includes a few more pieces that have already
moved to the providers folder. This will allow the providers folder to
only hold the list of providers in the future and the providers-sdk to
be used by all providers. This is merely the first step (and a big one
at that), so there are plenty more to come in the next 5 days.

Signed-off-by: Dominik Richter <dominik.richter@gmail.com>