🧹 decouple inventory manager from providers  (#1358)
Shoving the entire MQL runtime into providers is not very helpful, we
should keep them slim. This change moves the asset resolution with
credentials to the calling runtime (cnquery/cnspec) instead of the
plugin. This keeps plugins more focused and slim, fewer things for
developers to know.

get closer to turning the tests green

---------

Signed-off-by: Dominik Richter <dominik.richter@gmail.com>