🐛 Remove printer tests that fail
These tests started failing with #1346. The issue here is that code only
copies over things from the assessment block to the defaults block that
ultimately depend on the context. It ignores everything else as it would
not always be possible to merge the blocks. In this case, we could
technically merge those correctly, but the code as written assumes there
is only one root, the context. Since the queries as written don't make
sense, I'm removing them. I will come back and make changes to allow for
that case in another change