🐛 establish cache refresh from recording pattern (#1402)
recently introduced for services, this pattern is excellent to keep the
cache separate from the initial load (even if we have to cycle things
again) AND it makes it so that we can load these things safely from
recordings.

Up until this point there was the edge-case where we loaded a cached
list of resources from a recording during a call for eg
`resource(name)`. Since it is loaded from a recording, it did not create
the mqlResourceInternal structures. A nasty workaround to this problem
has now been removed as well. This model is much cleaner and safe to use
across recordings

Signed-off-by: Dominik Richter <dominik.richter@gmail.com>