✨ make providers/bundle
to generate tar.xz files for every bundle locally. Great for testing, not meant for distirbution (since you want to do this for every architecture)

Also: remove some older resource generation code, which we no longer use. Do not install the core provider by install (it is meant to be builtin by default). And finally: do not build the lr.manifest.json into the dist folder, since we don't distribute it. Instead it will be kept alongside the manifest.yaml for further processing.
Signed-off-by: Dominik Richter <dominik.richter@gmail.com>