🐛 pre-load all providers
The mechanisms where it was only loaded from disk when necessary breaks because we need the CLI options to be loaded before we de-dupe any providers. This is done via name, so we have to have their metadata loaded. This makes the code simpler and removes any surprises with having a provider that has no metadata.

Signed-off-by: Dominik Richter <dominik.richter@gmail.com>