Set the title of AWS accounts to be AWS Account
Right now an AWS Account passes into this functionw with the name of ""
and we set the name to "aws" and the title to "Amazon Web Services". The
"aws" platform is actually the account so we should properly name it.
This is important as time goes on with fine grained scanning where we
need to distinguish between the resources and the main account.

Signed-off-by: Tim Smith <tsmith84@gmail.com>