✨ migrate dns, openpgp, and parse.openpgp
As with certificates: `parse.openpgp` goes to `os` now, while `openpgp` moves to `network`. There is also a new resource: `openpgp.entities` which is private (because at the moment not user-facing, but this could change) and it takes the content of OpenPGP and then does all the parsing / entity creation.

Signed-off-by: Dominik Richter <dominik.richter@gmail.com>