🐛 Fix conversion of primitives with NO type
In case we have a resource without data, it also does not have a type.
We need to catch this to prevent the error: `cannot convert primitive with NO type information`

This happens, e.g., when we test for the exists of a file before doing anything else with the resource.

Partially-Fixes #957

Signed-off-by: Christian Zunker <christian@mondoo.com>