⭐️ Discovery for GCP resources (#1590)
This PR builds on top of #1587

This changes ports over the old mql based approach from v8 to v9.
Instead of using MQL, we use the resources directly. This new approach
has a lot of advantages compared to the previous one. We have full type
support when we use resources and we can directly use every field
defined in lr. On top the code is much simplified since we just have one
file where all the discovery happens.

---------

Signed-off-by: Ivan Milchev <ivan@mondoo.com>
Co-authored-by: Victoria Jeffrey <vj@mondoo.com>
Co-authored-by: Ivan Milchev <ivan@mondoo.com>