🐛 create LR before CLI config (#1624)
In the case of K8s, the config pulls in discovery options from the
remaining provider, which ultimately pull a dependency to the k8s
resources. These, however, do not exist on a clean build, because LR has
not yet been executed. Swap the order so that LR runs first, creating
the lr-generated files (like k8s.lr.go) which then make the dependency
chain work for creating the config.

Signed-off-by: Dominik Richter <dominik.richter@gmail.com>